+++ x/stake/keeper/test_common.go
@@ -110,11 +110,14 @@ func CreateTestInput(t *testing.T, isCheckTx bool, initCoins int64) (sdk.Context
 	keeper.SetPool(ctx, types.InitialPool())
 	keeper.SetNewParams(ctx, types.DefaultParams())
 
-	// fill all the addresses with some coins
+	// fill all the addresses with some coins, set the loose pool tokens simultaneously
 	for _, addr := range Addrs {
+		pool := keeper.GetPool(ctx)
 		ck.AddCoins(ctx, addr, sdk.Coins{
 			{keeper.GetParams(ctx).BondDenom, sdk.NewInt(initCoins)},
 		})
+		pool.LooseTokens += initCoins
+		keeper.SetPool(ctx, pool)
 	}
 
 	return ctx, accountMapper, keeper