+++ x/stake/types/pool.go
@@ -2,21 +2,22 @@ package types
 
 import (
 	"bytes"
+	"fmt"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 // Pool - dynamic parameters of the current state
 type Pool struct {
-	LooseUnbondedTokens int64   `json:"loose_unbonded_tokens"` // tokens not associated with any validator
-	UnbondedTokens      int64   `json:"unbonded_tokens"`       // reserve of unbonded tokens held with validators
-	UnbondingTokens     int64   `json:"unbonding_tokens"`      // tokens moving from bonded to unbonded pool
-	BondedTokens        int64   `json:"bonded_tokens"`         // reserve of bonded tokens
-	UnbondedShares      sdk.Rat `json:"unbonded_shares"`       // sum of all shares distributed for the Unbonded Pool
-	UnbondingShares     sdk.Rat `json:"unbonding_shares"`      // shares moving from Bonded to Unbonded Pool
-	BondedShares        sdk.Rat `json:"bonded_shares"`         // sum of all shares distributed for the Bonded Pool
-	InflationLastTime   int64   `json:"inflation_last_time"`   // block which the last inflation was processed // TODO make time
-	Inflation           sdk.Rat `json:"inflation"`             // current annual inflation rate
+	LooseTokens       int64   `json:"loose_unbonded_tokens"` // tokens not associated with any validator
+	UnbondedTokens    int64   `json:"unbonded_tokens"`       // reserve of unbonded tokens held with validators
+	UnbondingTokens   int64   `json:"unbonding_tokens"`      // tokens moving from bonded to unbonded pool
+	BondedTokens      int64   `json:"bonded_tokens"`         // reserve of bonded tokens
+	UnbondedShares    sdk.Rat `json:"unbonded_shares"`       // sum of all shares distributed for the Unbonded Pool
+	UnbondingShares   sdk.Rat `json:"unbonding_shares"`      // shares moving from Bonded to Unbonded Pool
+	BondedShares      sdk.Rat `json:"bonded_shares"`         // sum of all shares distributed for the Bonded Pool
+	InflationLastTime int64   `json:"inflation_last_time"`   // block which the last inflation was processed // TODO make time
+	Inflation         sdk.Rat `json:"inflation"`             // current annual inflation rate
 
 	DateLastCommissionReset int64 `json:"date_last_commission_reset"` // unix timestamp for last commission accounting reset (daily)
 
@@ -34,7 +35,7 @@ func (p Pool) Equal(p2 Pool) bool {
 // initial pool for testing
 func InitialPool() Pool {
 	return Pool{
-		LooseUnbondedTokens:     0,
+		LooseTokens:             0,
 		BondedTokens:            0,
 		UnbondingTokens:         0,
 		UnbondedTokens:          0,
@@ -52,7 +53,7 @@ func InitialPool() Pool {
 
 // Sum total of all staking tokens in the pool
 func (p Pool) TokenSupply() int64 {
-	return p.LooseUnbondedTokens + p.UnbondedTokens + p.UnbondingTokens + p.BondedTokens
+	return p.LooseTokens + p.UnbondedTokens + p.UnbondingTokens + p.BondedTokens
 }
 
 //____________________________________________________________________
@@ -95,6 +96,10 @@ func (p Pool) addTokensUnbonded(amount int64) (p2 Pool, issuedShares PoolShares)
 	issuedSharesAmount := sdk.NewRat(amount).Quo(p.UnbondedShareExRate()) // tokens * (shares/tokens)
 	p.UnbondedShares = p.UnbondedShares.Add(issuedSharesAmount)
 	p.UnbondedTokens += amount
+	p.LooseTokens -= amount
+	if p.LooseTokens < 0 {
+		panic(fmt.Sprintf("sanity check: loose tokens negative, pool: %v", p))
+	}
 	return p, NewUnbondedShares(issuedSharesAmount)
 }
 
@@ -102,6 +107,10 @@ func (p Pool) removeSharesUnbonded(shares sdk.Rat) (p2 Pool, removedTokens int64
 	removedTokens = p.UnbondedShareExRate().Mul(shares).Evaluate() // (tokens/shares) * shares
 	p.UnbondedShares = p.UnbondedShares.Sub(shares)
 	p.UnbondedTokens -= removedTokens
+	p.LooseTokens += removedTokens
+	if p.UnbondedTokens < 0 {
+		panic(fmt.Sprintf("sanity check: unbonded tokens negative, pool: %v", p))
+	}
 	return p, removedTokens
 }
 
@@ -109,6 +118,10 @@ func (p Pool) addTokensUnbonding(amount int64) (p2 Pool, issuedShares PoolShares
 	issuedSharesAmount := sdk.NewRat(amount).Quo(p.UnbondingShareExRate()) // tokens * (shares/tokens)
 	p.UnbondingShares = p.UnbondingShares.Add(issuedSharesAmount)
 	p.UnbondingTokens += amount
+	p.LooseTokens -= amount
+	if p.LooseTokens < 0 {
+		panic(fmt.Sprintf("sanity check: loose tokens negative, pool: %v", p))
+	}
 	return p, NewUnbondingShares(issuedSharesAmount)
 }
 
@@ -116,6 +129,10 @@ func (p Pool) removeSharesUnbonding(shares sdk.Rat) (p2 Pool, removedTokens int6
 	removedTokens = p.UnbondingShareExRate().Mul(shares).Evaluate() // (tokens/shares) * shares
 	p.UnbondingShares = p.UnbondingShares.Sub(shares)
 	p.UnbondingTokens -= removedTokens
+	p.LooseTokens += removedTokens
+	if p.UnbondedTokens < 0 {
+		panic(fmt.Sprintf("sanity check: unbonding tokens negative, pool: %v", p))
+	}
 	return p, removedTokens
 }
 
@@ -123,6 +140,10 @@ func (p Pool) addTokensBonded(amount int64) (p2 Pool, issuedShares PoolShares) {
 	issuedSharesAmount := sdk.NewRat(amount).Quo(p.BondedShareExRate()) // tokens * (shares/tokens)
 	p.BondedShares = p.BondedShares.Add(issuedSharesAmount)
 	p.BondedTokens += amount
+	p.LooseTokens -= amount
+	if p.LooseTokens < 0 {
+		panic(fmt.Sprintf("sanity check: loose tokens negative, pool: %v", p))
+	}
 	return p, NewBondedShares(issuedSharesAmount)
 }
 
@@ -130,5 +151,9 @@ func (p Pool) removeSharesBonded(shares sdk.Rat) (p2 Pool, removedTokens int64) 
 	removedTokens = p.BondedShareExRate().Mul(shares).Evaluate() // (tokens/shares) * shares
 	p.BondedShares = p.BondedShares.Sub(shares)
 	p.BondedTokens -= removedTokens
+	p.LooseTokens += removedTokens
+	if p.UnbondedTokens < 0 {
+		panic(fmt.Sprintf("sanity check: bonded tokens negative, pool: %v", p))
+	}
 	return p, removedTokens
 }